+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Changelog
 0.5.0 (2015-05-XX)
 ++++++++++++++++++
 
+- Added: Django: Support for multiple databases (`Issue #9 <https://github.com/maxtepkeev/architect
+  /issues/9>`__)
 - Changed: ``dsn`` partition option renamed to ``db`` to cover more use cases
 - Changed: ``DsnParseError`` renamed to ``OptionValueError`` to cover more use cases